[v0.0.1 🎉] fixed 7 build issues ✨
