--- .husky/pre-commit
@@ -1,1 +1,0 @@
-pnpm lint && pnpm build:check && pnpm format
