+++ changelog.md
@@ -1,0 +1,8 @@
+### Sabbir Branch Changelog [v0.0.1]
+
+- [x] `i18n` language support issue fixed
+- [x] Removed husky from our project
+- [x] Swiper css module not found issue fixed
+- [x] Local language not changing issue [getRequstConfig](https://next-intl.dev/blog/next-intl-3-22#await-request-locale)
+- [x] Change create middleware function defination using routing
+- [x] Fixed Language Switcher issues
