+++ src/middleware.ts
@@ -1,13 +1,7 @@
 import createMiddleware from 'next-intl/middleware';
+import  {routing} from '@/routing'
 
-import { defaultLocale, localePrefix, locales, pathnames } from './i18n';
-
-export default createMiddleware({
-  defaultLocale,
-  locales,
-  localePrefix,
-  pathnames,
-});
+export default createMiddleware(routing);
 
 export const config = {
   matcher: [
